requirements.txt
Django==4.2.28
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.2.0
psycopg[binary]==3.2.9
openai==1.79.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
scikit-learn==1.5.1
joblib==1.4.2
numpy==1.26.4
pypdf==5.1.0
python-docx==1.1.2